fixed bugs on local issues page, search bar is now working on both tabs
