+++ client/src/components/LocalIssues.jsx
@@ -3,12 +3,13 @@ import { AppContext } from '../context/AppContext';
 import Cards from './Card';
 
 const LocalIssues = () => {
-  const { apiData, setApiData } = useContext(AppContext);
+  const { apiData, setApiData, filterData } = useContext(AppContext);
 
   useEffect(() => {
     fetch(`/localissuespage`)
       .then((res) => res.json())
       .then((data) => {
+        let mapData = data.slice(1);
         setApiData(data.slice(1));
         console.log(data);
       });
@@ -16,19 +17,20 @@ const LocalIssues = () => {
 
   return (
     <div className="test">
-      {apiData.map(
-        (item) =>
-          Object.keys(item).length && (
-            <Cards
-              city={item.city.replaceAll('_', ' ')}
-              issueType={item.issue_type}
-              streetAddress={item.street_address}
-              zipCode={item.zip_code}
-              locationLat={item.location ? item.location.latitude : 0}
-              locationLng={item.location ? item.location.longitude : 0}
-            />
-          )
-      )}
+      {filterData &&
+        filterData.map(
+          (item) =>
+            Object.keys(item).length && (
+              <Cards
+                city={item.city.replaceAll('_', ' ')}
+                issueType={item.issue_type}
+                streetAddress={item.street_address}
+                zipCode={item.zip_code}
+                locationLat={item.location ? item.location.latitude : 0}
+                locationLng={item.location ? item.location.longitude : 0}
+              />
+            )
+        )}
     </div>
   );
 };
