+++ client/src/pages/LocalVoice/LocalVoice.css
@@ -42,5 +42,5 @@
   color: white;
   line-height: 0.5rem;
   margin-top: 1rem;
-  background-image: url('../../assets/images/submitBkgdImage.svg');
+  /* background-image: url('../../assets/images/submitBkgdImage.svg'); */
 }
